requirements.txt
streamlit
mysql.connector
contextlib
pandas
plotly.express